third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-preventdefault.html

<!DOCTYPE HTML>
<script src="../../../resources/js-test.js"></script>
<style>
div.box {
  margin: 10px;
  padding: 50px;
  float: left;
}
</style>

<div id="target" class="box" style="background-color:green;">
</div>

<div id="console"></div>

<script>
description("Verifies the effect of pointer event prevent-default on following pointer & mouse events.");

// TODO(mustaq): Grow the lists as we add support for more PEs.
var mouseEventList = ["mousedown", "mouseup", "mouseenter", "mouseleave", "mouseover", "mouseout"];
var pointerEventList = ["pointerdown", "pointerup", "pointerenter", "pointerleave", "pointerover", "pointerout"];

var preventDefaultList = ["none"].concat(pointerEventList);
var eventToPreventDefault = "";
var seqNo = 0;

function init() {
  var targetDiv = document.getElementById("target");

  mouseEventList.forEach(function(eventName) {
    targetDiv.addEventListener(eventName, function(event) {
      debug(++seqNo + ". target received " + event.type);
    });
  });

  pointerEventList.forEach(function(eventName) {
    targetDiv.addEventListener(eventName, function(event) {
      var preventDefaultMsg = "";
      if (event.type == eventToPreventDefault) {
        event.preventDefault();
        preventDefaultMsg = ", prevented default";
      }

      debug(++seqNo + ". target received " + event.type + preventDefaultMsg);
    });
  });
}

function runTests() {
  var rect = document.getElementById("target").getBoundingClientRect();
  var x1 = rect.left - 3;
  var y1 = rect.top - 5;
  var x2 = rect.left + 3;
  var y2 = rect.top + 5;

  preventDefaultList.forEach(function(preventDefaultEvent) {
    seqNo = 0;

    debug("--- test with preventDefault on " + preventDefaultEvent + " ---");
    eventToPreventDefault = preventDefaultEvent;

    debug("- start with mouse outside target -");
    eventSender.mouseMoveTo(x1, y1);

    debug("- move into target & click -");
    eventSender.mouseMoveTo(x2, y2);
    eventSender.mouseDown();
    eventSender.mouseUp();

    debug("- drag & release within target -");
    eventSender.mouseDown();
    eventSender.mouseMoveTo(x2+5, y2+5);
    eventSender.mouseUp();

    // TODO(mustaq): Add drag across target border after fixing crbug.com/356090.

    debug("- move outside target again -");
    eventSender.mouseMoveTo(x1, y1);

    debug("");
  });
}

init();
if (window.eventSender)
  runTests();
else
  debug("This test requires eventSender");

</script>
